+++ CHANGELOG.md
@@ -6,6 +6,7 @@ alpha
   * update to bootstrap-helper 2.1.2.1
   * remove newrelic_rpm
   * rvm-capistrano
+  * replace devise.email_sender to Setting.email_sender
   
 0.1.0 (Nov 12, 2012)
   * add powder to Gemfile